Refactor `thewalrus.quantum` (#197)
* Refactor thewalrus.quantum

* Fix imports

* Remove old quantum.py

* Fix import

* Fix pylint issues

* Add import

* Add find_packages to setup

* Apply suggestions from code review


* Finalise refactor

* Fix spelling

* Updates from code-review

* Update doc

* Add deprecation warning

* Fix pylint

* Use `functools.wrap`


* Update thewalrus/quantum/__init__.py


* Import functools
